move services to new unified folder, god mode enables armor and barrier too
