+++ dll/src/cheats/Core/CheatMain.cpp
@@ -1,6 +1,6 @@
 #include "CheatMain.h"
 #include "Config.h"
-#include "../Features/WeaponService.h"
+#include "../Services/WeaponService.h"
 #include "../Services/AimbotService.h"
 #include "../Services/BoneService.h"
 #include "../Utils/Console.h"
@@ -162,7 +162,7 @@ namespace Cheat {
             Services::BoneService::Initialize();
 
             // Initialize weapon service
-            Features::WeaponService::Initialize();
+            Services::WeaponService::Initialize();
 
             return true;
         }
@@ -173,24 +173,7 @@ namespace Cheat {
             if (Utils::Input::IsKeyPressed(Config::Hotkeys::DumpBones)) {
                 Services::BoneService::DumpUniqueEnemyBones(Cheat::Services::GameServices::GetWorld());
                 // Services::BoneService::DumpPlayerBones(Cheat::Services::GameServices::GetRPlayerPawn());
-                auto healthComponent = Cheat::Services::GameServices::GetRPlayerPawn()->GetHealthComponent();
-                /** Dump the invulnerable state of the player. Currentvalue, value, absolute, and range
-                 * struct alignas(0x08) FRMutableInteger final
-{
-public:
-	int32                                         CurrentValue;                                      // 0x0000(0x0004)(BlueprintVisible, BlueprintReadOnly, ZeroConstructor, IsPlainOldData, RepSkip, NoDestructor, HasGetValueTypeHash, NativeAccessSpecifierPublic)
-	int32                                         Value;                                             // 0x0004(0x0004)(Edit, BlueprintVisible, ZeroConstructor, DisableEditOnInstance, IsPlainOldData, NoDestructor, HasGetValueTypeHash, NativeAccessSpecifierPublic)
-	int32                                         AbsoluteValue;                                     // 0x0008(0x0004)(BlueprintVisible, BlueprintReadOnly, ZeroConstructor, IsPlainOldData, RepSkip, NoDestructor, HasGetValueTypeHash, NativeAccessSpecifierPublic)
-	struct FIntPoint                              MinMaxRange;                                       // 0x000C(0x0008)(Edit, BlueprintVisible, BlueprintReadOnly, ZeroConstructor, DisableEditOnInstance, IsPlainOldData, RepSkip, NoDestructor, HasGetValueTypeHash, NativeAccessSpecifierPublic)
-	uint8                                         Pad_14[0x10C];                                     // 0x0014(0x010C)(Fixing Struct Size After Last Property [ Dumper-7 ])
-};
-                 */
-SDK::FRMutableInteger invuln = healthComponent->Invulnerable;
-LOG_INFO("Invulnerable: %d", invuln.CurrentValue);
-LOG_INFO("Invulnerable: %d", invuln.Value);
-LOG_INFO("Invulnerable: %d", invuln.AbsoluteValue);
-LOG_INFO("Invulnerable: %d", invuln.MinMaxRange.X);
-LOG_INFO("Invulnerable: %d", invuln.MinMaxRange.Y);
+               
 
 
             }
@@ -202,7 +185,7 @@ LOG_INFO("Invulnerable: %d", invuln.MinMaxRange.Y);
 
             // Handle log weapon stats key
             if (Utils::Input::IsKeyPressed(Config::Hotkeys::LogWeaponStats)) {
-                Features::WeaponService::LogAllWeaponStats();
+                Services::WeaponService::LogAllWeaponStats();
             }
         }
 
@@ -211,7 +194,7 @@ LOG_INFO("Invulnerable: %d", invuln.MinMaxRange.Y);
             Cheat::Services::PlayerEffectsService::Update(Cheat::Services::GameServices::GetPlayerController());
 
             // Update weapon service
-            Features::WeaponService::Update();
+            Services::WeaponService::Update();
 
             // Update aimbot system
             Services::AimbotService::Update(deltaTime);
