+++ dll/src/gui/Tabs/WeaponsTab.cpp
@@ -1,7 +1,7 @@
 #include "WeaponsTab.h"
 #include <imgui.h>
 #include "../../cheats/Core/Config.h"
-#include "../../cheats/Features/WeaponService.h"
+#include "../../cheats/Services/WeaponService.h"
 
 namespace CheatMenu { namespace Tabs {
 
@@ -10,24 +10,24 @@ void WeaponsTab() {
     ImGui::Separator();
 
     if (ImGui::Checkbox("Infinite Ammo", &Cheat::Config::Features::InfiniteAmmo)) {
-        Cheat::Features::WeaponService::OnWeaponSettingsChanged();
+        Cheat::Services::WeaponService::OnWeaponSettingsChanged();
     }
     if (ImGui::IsItemHovered()) ImGui::SetTooltip("No ammo cost when firing");
 
     if (ImGui::Checkbox("Increased Damage", &Cheat::Config::Features::IncreasedDamage)) {
-        Cheat::Features::WeaponService::OnWeaponSettingsChanged();
+        Cheat::Services::WeaponService::OnWeaponSettingsChanged();
     }
     if (ImGui::IsItemHovered()) ImGui::SetTooltip("Use slider to apply a damage multiplier");
     if (Cheat::Config::Features::IncreasedDamage) {
         ImGui::Indent();
         if (ImGui::SliderFloat("Damage Multiplier", &Cheat::Config::Features::DamageMultiplier, 0.1f, 10.0f, "x%.2f")) {
-            Cheat::Features::WeaponService::OnWeaponSettingsChanged();
+            Cheat::Services::WeaponService::OnWeaponSettingsChanged();
         }
         ImGui::Unindent();
     }
 
     if (ImGui::Checkbox("Crit Multiplier", &Cheat::Config::Features::HighCritMultiplier)) {
-        Cheat::Features::WeaponService::OnWeaponSettingsChanged();
+        Cheat::Services::WeaponService::OnWeaponSettingsChanged();
     }
     if (ImGui::IsItemHovered()) ImGui::SetTooltip("Use slider to apply a crit multiplier");
     if (Cheat::Config::Features::HighCritMultiplier) {
@@ -35,31 +35,31 @@ void WeaponsTab() {
         // Use unique ID scope to avoid label collisions with other sliders named similarly.
         ImGui::PushID("CritMultiplierSlider");
         if (ImGui::SliderFloat("Crit Multiplier Value", &Cheat::Config::Features::CritMultiplier, 0.1f, 10.0f, "x%.2f")) {
-            Cheat::Features::WeaponService::OnWeaponSettingsChanged();
+            Cheat::Services::WeaponService::OnWeaponSettingsChanged();
         }
         ImGui::PopID();
         ImGui::Unindent();
     }
 
     if (ImGui::Checkbox("Rate of Fire", &Cheat::Config::Features::RateOfFireOverride)) {
-        Cheat::Features::WeaponService::OnWeaponSettingsChanged();
+        Cheat::Services::WeaponService::OnWeaponSettingsChanged();
     }
     if (ImGui::IsItemHovered()) ImGui::SetTooltip("Override base rate of fire using the slider below");
     if (Cheat::Config::Features::RateOfFireOverride) {
         ImGui::Indent();
         if (ImGui::SliderFloat("Rate of Fire Value", &Cheat::Config::Features::RateOfFireValue, 0.1f, 100.0f)) {
-            Cheat::Features::WeaponService::OnWeaponSettingsChanged();
+            Cheat::Services::WeaponService::OnWeaponSettingsChanged();
         }
         ImGui::Unindent();
     }
 
     if (ImGui::Checkbox("No Cooldown", &Cheat::Config::Features::NoCooldown)) {
-        Cheat::Features::WeaponService::OnWeaponSettingsChanged();
+        Cheat::Services::WeaponService::OnWeaponSettingsChanged();
     }
     if (ImGui::IsItemHovered()) ImGui::SetTooltip("No weapon cooldown");
 
     if (ImGui::Checkbox("No Recoil", &Cheat::Config::Features::NoRecoil)) {
-        Cheat::Features::WeaponService::OnWeaponSettingsChanged();
+        Cheat::Services::WeaponService::OnWeaponSettingsChanged();
     }
     if (ImGui::IsItemHovered()) ImGui::SetTooltip("No recoil + no spread");
